Created tests for scrape_weather_data method.
Created the test_scrape_weather_data test.  This tests to see if when given a valid url which we already know what the response will be, the
scrape_weather_data function outputs the correct data in json format.
Created the test_scrape_weather_data_404 test.  This tests the scrape_weather_data method to make sure that if we give it an address which
leads to a 404 error, it handles this gracefully.
Created the test_scrape_weather_data_no_table function.  This makes sure that scrape_weather_data doesn't try to get a table from any random
webpage which we were able to go to.