Created validate_year and tests.
validate_year ensures that the user provided year is a valid year, given a validated day and month.  Year cannot be in the future, must have the correct number of digits, and it also
handles situations where it is a leap year and checks if the day provided for February.
Created tests to check:
a valid month/day/year is true
false if it isn't a leap year and user provided feb 29th
year string is too many digits
year string is too few digits
Year/Month/Day is current or in the past.