Created validate_location and tests.
validate_location validates a user inputted location string as either a 'city, state', zipcode, or airport code.  If it doesn't fall
under the formatting of one of those three, the function which calls it, get_inputs, prints that there was an error with the location
provided by the user and exits.