Builds json output using json module. Removed old helper function and tests.
created new get_cell_data helper function to cleanly get data from a cell and return it.
Now stores temperature data in a dictionary and uses the json module to convert that to a json string when the function returns.
deleted the helper function that was used to help build the json string before.
Modified tests because the keys are now sorted and removes the degree character.  Tests now expect sorted keys in the json string.
Closes #13