Modified scrape_weather_data.  Created table_is_complete.
scrape_weather_data now calls a helper method, table_is_complete, to verify that the table we received has all the headers required to get all the data
that we want.  Some pages have the table but do not have all the data asked for.
Prints an error explaining that the table is incomplete in this situation.
table_is_complete accepts a soup object with table rows in it.  Scans for each header we want and only returns true if it found all of them.
Created tests to verify that it works for html with the headers we want and doesn't work if all the headers aren't in the html.
Closes #16